Allow setting custom FQDN
During the interactive installation it is possible to specify
a custom FQDN instead of using an autodetected host name. This is
especially useful when using virtual hosts on the server hosting
oVirt Engine.

This patch allows setting ovirt_engine_setup_fqdn variable which
will be used during the installation process to set the value to
a desired one.

Closes-bug: #49