Add ovirt_engine_setup_apache_ssl variable
When we want to use httpd certificates signed by other CA, we need to set the
OVESETUP_APACHE_CONFIG_SSL variable to ‘manual’. There's no need to supply
all of the answer file manually though so let's have a role variable for
setting that. By default, it's not defined so engine's default (‘automatic’)
will be used.